% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_shadowr.R
\docType{class}
\name{shadow-class}
\alias{shadow-class}
\alias{find_element}
\alias{find_element,shadow,character,missing-method}
\alias{find_element,shadow,character,webElement-method}
\alias{find_elements}
\alias{find_elements,shadow,character,missing-method}
\alias{find_elements,shadow,character,webElement-method}
\alias{get_shadow_element}
\alias{get_shadow_element,shadow,webElement,character-method}
\alias{get_all_shadow_element}
\alias{get_all_shadow_element,shadow,webElement,character-method}
\alias{get_parent_element}
\alias{get_parent_element,shadow,webElement-method}
\alias{get_child_elements}
\alias{get_child_elements,shadow,webElement-method}
\alias{get_sibling_element}
\alias{get_sibling_element,shadow,webElement,character-method}
\alias{get_previous_sibling_element}
\alias{get_previous_sibling_element,shadow,webElement-method}
\alias{get_next_sibling_element}
\alias{get_next_sibling_element,shadow,webElement-method}
\alias{scroll_to}
\alias{scroll_to,shadow,webElement-method}
\alias{is_checked}
\alias{is_checked,shadow,webElement-method}
\alias{is_disabled}
\alias{is_disabled,shadow,webElement-method}
\alias{is_visible}
\alias{is_visible,shadow,webElement-method}
\title{CLASS shadow}
\usage{
\dontshow{}
The following methods can be used with the shadow Class

\dontshow{}
\strong{find_element}(css_selector) :
Use this method if want single element from DOM
\cr

\dontshow{}
\strong{find_element}(object parent, str css_selector):
Use this if you want to find a single elements from parent object DOM

\dontshow{}

\dontshow{}
\strong{find_elements}(str css_selector):
Use this if you want to find all elements from DOM
\cr

\dontshow{}
\strong{find_elements}(object parent, str css_selector):
Use this if you want to find all elements from parent object DOM

\dontshow{}

\dontshow{}
\strong{get_shadow_element}(object parent,str css_selector):
Use this if you want to find a single element from parent DOM

\dontshow{}

\dontshow{}
\strong{get_all_shadow_element}(object parent, str css_selector):
Use this if you want to find all elements from parent DOM

\dontshow{}

\dontshow{}
\strong{get_parent_element}(object element):
Use this to get the parent element if web element

\dontshow{}

\dontshow{}
\strong{get_child_elements}(object parent) :
Use this to get all the child elements of parent element

\dontshow{}

\dontshow{}
\strong{get_sibling_element}(object element, str css_selector):
Use this to get adjacent(sibling) element using css selector

\dontshow{}

\dontshow{}
\strong{get_previous_sibling_element}(object element):
Use this to get previous adjacent(sibling) element

\dontshow{}

\dontshow{}
\strong{get_next_sibling_element}(object element):
Use this to get next adjacent(sibling) element

\dontshow{}

\dontshow{}
\strong{scroll_to}(object element):
Use this to scroll to web element

\dontshow{}

\dontshow{}
\strong{is_checked}(object element):
Use this if you want to check if checkbox is selected

\dontshow{}

\dontshow{}
\strong{is_disabled}(object element):
Use this if you want to check if element is disabled

\dontshow{}

\dontshow{}
\strong{is_visible}(object element):
Use this if you want to find visibility of element
}
\arguments{
\item{theObject}{the shadow class}

\item{css_selector}{a css string to find a webelement}

\item{element}{a RSelenium webelement}
}
\description{
Selenium plugin to manage multi level shadow DOM elements on web page.
}
\section{Slots}{

\describe{
\item{\code{driver}}{The shadow class takes a RSelenium RemoteDriver}

\item{\code{javascript_library}}{the selenium plugin}
}}

\seealso{
\url{https://github.com/Ricardo281/shadowr} for examples
}
\keyword{shadow}
