% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowr.R
\name{shadow}
\alias{shadow}
\title{Shadow Class Constructer}
\usage{
shadow(driver)
}
\arguments{
\item{driver}{The shadow class takes a RSelenium RemoteDriver}
}
\description{
This is a Class Constructer funtion to create a shadow class and are able to use the
Selenium plugin to manage multi level shadow DOM elements on a web page.
}
\seealso{
\code{\link{shadow-class}} or
\cr
\url{https://github.com/Ricardo281/shadowr} for examples
}
\keyword{shadow}
